// front/src/data/types/events.ts
export interface EventsInterface {
  users: any
  event: any
  idEvents: number
  id: number
  title: string
  description: string
  price: number
  department: string
  occupation: string
  statusEvent: number
  idFile: number
  dateEvent: string
  timeEvent: string
  qtdParcelamento: number
  pathName: string
  idHasEvents: number
  name: string
}